fastapi/pyproject.toml
[project]
name = "ai-interview-agent-fastapi"
version = "1.0.0"
description = "AI Interview Agent - FastAPI Implementation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic[email]",
    "pydantic-settings",
    "python-multipart",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-dotenv",
    "redis",
    "celery",
    "kombu",
    "websockets",
    "simple-websocket",
    "httpx",
    "aiohttp",
    "scikit-learn",
    "implicit",
    "faiss-cpu",
    "sentence-transformers",
    "pymilvus>=2.4.0",
    "onnxruntime==1.16.3",
    "numpy",
    "pandas",
    "jieba",
    "psycopg2>=2.9.11",
    "scipy>=1.15.3",
    "dashscope>=1.25.12",
    "structlog>=24.1.0",
    "python-json-logger>=2.0.7",
    "oss2>=2.18.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "setuptools<70",
]
